// prisma/generated/type-graphql/resolvers/inputs/Plugin_translationWhereUniqueInput.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../../scalars";
import { BytesFilter } from "../inputs/BytesFilter";
import { DateTimeFilter } from "../inputs/DateTimeFilter";
import { DateTimeNullableFilter } from "../inputs/DateTimeNullableFilter";
import { LanguageRelationFilter } from "../inputs/LanguageRelationFilter";
import { PluginRelationFilter } from "../inputs/PluginRelationFilter";
import { Plugin_translationWhereInput } from "../inputs/Plugin_translationWhereInput";
import { StringNullableFilter } from "../inputs/StringNullableFilter";
import { plugin_translationPlugin_idLanguage_idCompoundUniqueInput } from "../inputs/plugin_translationPlugin_idLanguage_idCompoundUniqueInput";

@TypeGraphQL.InputType("Plugin_translationWhereUniqueInput", {})
export class Plugin_translationWhereUniqueInput {
  @TypeGraphQL.Field(_type => plugin_translationPlugin_idLanguage_idCompoundUniqueInput, {
    nullable: true
  })
  plugin_id_language_id?: plugin_translationPlugin_idLanguage_idCompoundUniqueInput | undefined;

  @TypeGraphQL.Field(_type => [Plugin_translationWhereInput], {
    nullable: true
  })
  AND?: Plugin_translationWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [Plugin_translationWhereInput], {
    nullable: true
  })
  OR?: Plugin_translationWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [Plugin_translationWhereInput], {
    nullable: true
  })
  NOT?: Plugin_translationWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => BytesFilter, {
    nullable: true
  })
  plugin_id?: BytesFilter | undefined;

  @TypeGraphQL.Field(_type => BytesFilter, {
    nullable: true
  })
  language_id?: BytesFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  label?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  description?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  manufacturer_link?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  support_link?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  changelog?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  custom_fields?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  created_at?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeNullableFilter, {
    nullable: true
  })
  updated_at?: DateTimeNullableFilter | undefined;

  @TypeGraphQL.Field(_type => LanguageRelationFilter, {
    nullable: true
  })
  language?: LanguageRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PluginRelationFilter, {
    nullable: true
  })
  plugin?: PluginRelationFilter | undefined;
}